Moved classes into packages for organization
