--- src/main/java/Tagger.java
@@ -1,13 +1,0 @@
-public class Tagger implements InteractionManager{
-
-    private String tag;
-
-    @Override
-    public void interact() {
-
-    }
-
-    public void addTag(String tag){
-
-    }
-}
