--- src/main/java/Zoomer.java
@@ -1,13 +1,0 @@
-public class Zoomer implements InteractionManager{
-
-    private Object toZoom;
-
-    @Override
-    public void interact() {
-
-    }
-
-    public void zoom(Object toZoom){
-
-    }
-}
