+++ src/main/java/app/draweables/SelectionTool.java
@@ -1,8 +1,11 @@
+package app.draweables;
+
+import app.draweables.Drawable;
 import javafx.geometry.Point2D;
 
 public class SelectionTool implements Drawable {
 
-    /** Takes in two positions from mouse input and sends the area to the Zoomer
+    /** Takes in two positions from mouse input and sends the area to the app.interaction_managers.Zoomer
      */
     @Override
     public void useTwoPositions(Point2D first, Point2D second, String shape){
