+++ src/main/java/app/interaction_managers/Tagger.java
@@ -1,0 +1,17 @@
+package app.interaction_managers;
+
+import app.interaction_managers.InteractionManager;
+
+public class Tagger implements InteractionManager {
+
+    private String tag;
+
+    @Override
+    public void interact() {
+
+    }
+
+    public void addTag(String tag){
+
+    }
+}
