+++ src/main/java/app/media_managers/ImageModifier.java
@@ -1,11 +1,7 @@
-public class ImageModifier implements MediaManager{
+package app.media_managers;
+
+public class ImageModifier implements MediaManager {
     private String caption;
-
-    @Override
-    public void createObject() {
-
-    }
-
     @Override
     public void addMedia() {
