+++ src/main/java/storage/FileLoaderWriter.java
@@ -1,3 +1,7 @@
+package storage;
+
+import storage.Storage;
+
 public class FileLoaderWriter implements Storage {
     // TODO: Implement a file cache, which stores recently opened files to prevent consecutive reopening
     private int cacheSize = 10; // Size of the internal file cache
