+++ src/main/java/storage/Storage.java
@@ -1,3 +1,5 @@
+package storage;
+
 public interface Storage {
     public byte[] readFile(String path);
     public void writeFile(String path, byte[] Data);
